User implicit model binding and scopes

+++ database/seeders/DatabaseSeeder.php
@@ -8,33 +8,36 @@ use App\Models\Product;
 use App\Models\Transaction;
 use App\Models\User;
 use Illuminate\Database\Seeder;
+use Illuminate\Support\Facades\App;
 use Illuminate\Support\Facades\DB;
 
 class DatabaseSeeder extends Seeder
 {
     public function run()
     {
-        DB::statement('SET FOREIGN_KEY_CHECKS = 0');
+        if (App::environment('testing')) {
+            DB::statement('SET FOREIGN_KEY_CHECKS = 0');
 
-        User::truncate();
-        Category::truncate();
-        Product::truncate();
-        Transaction::truncate();
-        DB::table('category_product')->truncate();
+            User::truncate();
+            Category::truncate();
+            Product::truncate();
+            Transaction::truncate();
+            DB::table('category_product')->truncate();
 
-        $userQuantity = 50;
-        $categoryQuantity = 50;
-        $productQuantity = 50;
-        $transactionQuantity = 50;
+            $userQuantity = 5;
+            $categoryQuantity = 5;
+            $productQuantity = 5;
+            $transactionQuantity = 5;
 
-        User::factory()->count($userQuantity)->create();
-        Category::factory()->count($categoryQuantity)->create();
-        Product::factory()->count($productQuantity)->create()->each(function ($product) {
-            $categories = Category::all()->random(mt_rand(1, 5))->pluck('id');
-            $product->categories()->attach($categories);
-        });
-        Transaction::factory()->count($transactionQuantity)->create()->each(function ($transaction) {
+            User::factory()->count($userQuantity)->create();
+            Category::factory()->count($categoryQuantity)->create();
+            Product::factory()->count($productQuantity)->create()->each(function ($product) {
+                $categories = Category::all()->random(mt_rand(1, 5))->pluck('id');
+                $product->categories()->attach($categories);
+            });
+            Transaction::factory()->count($transactionQuantity)->create()->each(function ($transaction) {
 
-        });
+            });
+        }
     }
 }
